Revert "Revert "explicit skip for specs""
This reverts commit 97e3b485270bd4f0cdef64bb4b679909423a21d6.

+++ qa-tests-backend/src/test/groovy/AuditScrubbingTest.groovy
@@ -24,6 +24,9 @@ class AuditScrubbingTest extends BaseSpecification {
     private Notifier notifier
 
     def setupSpec() {
+        if (!shouldSpecRun(this.class.getSimpleName())) {
+            return
+        }
         def notifierConfig = NotifierService.getWebhookIntegrationConfig(
                 "audit-${getClass().name}-${UUID.randomUUID()}", false, "", true, true)
         notifier = NotifierService.addNotifier(notifierConfig)
@@ -101,6 +104,9 @@ class AuditScrubbingTest extends BaseSpecification {
     }
 
     def cleanupSpec() {
+        if (!shouldSpecRun(this.class.getSimpleName())) {
+            return
+        }
         if (notifier?.id) {
             NotifierService.deleteNotifier(notifier.id)
         }
